feat: start working on enacting proposals following ratification.
  The core idea is to stash store updates as we go, while mutating the
  RatificationContext we operate in. The data in the store that we use
  can in fact only be changed following those ratification updates; so
  we need not any kind of re-synchronisation and can apply updates
  directly on top of the snapshot we operate from.

  Because we need to immutability traverse the forest, while also
  occasionally allowing local mutations; we introduce a `Compass`, that
  acts as mutable iterator but without holding any reference on the
  forest (so it needs the forest each time it needs to access the next
  element).

  This allows to traverse the forest while remaining capable of
  borrowing an immutable reference to it on enactment.

  Now, we must also makes sure that the forest is getting properly
  pruned; which involves multiple steps:

  - Any impacted sub-tree must promote the enacted proposal as its new
    root, and prune all its siblings and their children. Those proposals
    are now unreachable / unvotable, and will be cleared from the state.

  - Any of the pruned proposal must be removed from the forest's lookup
    table;

  - The proposal sequence must be amended accordingly;

  - And finally, in case where a:

    - `hard_fork`; or
    - `constitutional_committee`; or
    - `constitution`

    proposal is ratified; then no other proposal can be ratified in the
    same epoch boundary. So any new cursor on the forest will yield
    `None`.

+++ crates/amaru-ledger/src/governance/ratification/constitutional_committee.rs
@@ -67,7 +67,7 @@ impl ConstitutionalCommittee {
         &self,
         current_epoch: Epoch,
         protocol_version: ProtocolVersion,
-        min_committee_size: usize,
+        min_committee_size: u16,
         proposal: &ProposalEnum,
     ) -> Option<&Rational64> {
         match proposal {
@@ -82,7 +82,7 @@ impl ConstitutionalCommittee {
 
                 // The minimum committee size has no effect during the bootstrap phase (i.e. v9). The
                 // committee is always allowed to vote during v9.
-                if active_members.len() < min_committee_size
+                if active_members.len() < (min_committee_size as usize)
                     && protocol_version > PROTOCOL_VERSION_9
                 {
                     return None;
